+++ src/database/operations/permissionGrants/loadPermissionGrantBundle.ts
@@ -1,0 +1,9 @@
+import { generateLoadBundleOperation } from '../generators';
+import type { PermissionGrant } from '../../../types';
+
+const loadPermissionGrantBundle = generateLoadBundleOperation<
+	PermissionGrant,
+	[]
+>('permissionGrants.selectWithPagination', 'permission_grants', []);
+
+export { loadPermissionGrantBundle };