+++ src/routers/usersRouter.ts
@@ -1,5 +1,8 @@
 import express from 'express';
 import { userChangemakerPermissionsHandlers } from '../handlers/userChangemakerPermissionsHandlers';
+import { userDataProviderPermissionsHandlers } from '../handlers/userDataProviderPermissionsHandlers';
+import { userFunderPermissionsHandlers } from '../handlers/userFunderPermissionsHandlers';
+import { userOpportunityPermissionsHandlers } from '../handlers/userOpportunityPermissionsHandlers';
 import { usersHandlers } from '../handlers/usersHandlers';
 import {
 	requireAdministratorRole,
@@ -9,9 +12,6 @@ import {
 	requireFunderPermission,
 } from '../middleware';
 import { Permission } from '../types';
-import { userFunderPermissionsHandlers } from '../handlers/userFunderPermissionsHandlers';
-import { userDataProviderPermissionsHandlers } from '../handlers/userDataProviderPermissionsHandlers';
-import { userOpportunityPermissionsHandlers } from '../handlers/userOpportunityPermissionsHandlers';
 
 const usersRouter = express.Router();